requirements.txt
requests
Pillow
tqdm
colorama
pyfiglet